omniparser-service/requirements.txt
flask==2.0.1
pillow==9.0.0
requests==2.26.0
gunicorn==20.1.0
